+++ resto-code-front/src/Components/TablaUsers/TablaUsuarios.jsx
@@ -22,13 +22,33 @@ const TablaUsuarios = () => {
   };
 
   const handleDelete = (id) => {
-    deleteUser(id);
     Swal.fire({
-      icon: "success",
-      title: "Usuario Eliminado",
-      showConfirmButton: false,
-      timer: 1500,
-    });
+      icon: 'warning',
+      title: 'Eliminar Usuario?',
+      showCancelButton: true,
+      confirmButtonText: 'Aceptar', 
+      cancelButtonText: 'Cancelar',
+      confirmButtonColor: '#651F71',
+      cancelButtonColor: '#C73333',
+      background: '#31302F',
+      color: 'white',
+      backdrop: `rgba(0,0,14,0.4)` 
+    }).then((result) => {        
+      if (result.isConfirmed) {
+        deleteUser(id);
+        Swal.fire({
+          icon: "success",
+          title: "Usuario Eliminado",
+          showConfirmButton: false,
+          timer: 1500,
+          background: '#31302F',
+          color: 'white',
+          backdrop: `rgba(0,0,14,0.4)` 
+        });
+      } else if (result.isDenied) {
+        return
+      }
+    })    
   };
 
   return (
@@ -37,7 +57,7 @@ const TablaUsuarios = () => {
         <h1>No hay usuarios registrados</h1>
       ) : (
         <>
-          <Table responsive className="table-dark table-hover text-center">
+          <Table responsive className="table-dark table-hover text-center tabla">
             <thead>
               <tr>
                 <th>Nombre</th>
@@ -50,11 +70,11 @@ const TablaUsuarios = () => {
             <tbody>
               {users.map((user) => (
                 <tr key={user.id}>
-                  <td>{user.nombre}</td>
-                  <td>{user.apellido}</td>
-                  <td>{user.edad}</td>
-                  <td>{user.email}</td>
-                  <td>
+                  <td data-label="Nombre">{user.nombre}</td>
+                  <td data-label="Apellido">{user.apellido}</td>
+                  <td data-label="Edad">{user.edad}</td>
+                  <td data-label="Email">{user.email}</td>
+                  <td data-label="Acciones">
                     <button
                       className="botonEditar m-1"
                       onClick={() => handleEdit(user)}
